refactor: Refactor common packages for readability
