--- lambda_backend/user_microservice/src/common/index.ts
@@ -1,4 +1,0 @@
-export * from './types';
-export * from './schema';
-export * from './middleware';
-export * from './identityProvider';
